Introduce GenericPacket<N> and GenericPacketBatch<N>
These new items are like their "regular" predecessors, except with the
addition of a const generic parameter N. In this context, N represents
the length of the buffer in GenericPacket, in bytes.

Type aliases Packet and PacketBatch have been added over the generic
types with N = PACKET_DATA_SIZE. This updates all callers of the old
structures automatically.