Enough to get ledger-tool compiling on Linux
Rocksdb didn't want to compile as-is, so built locally and adjusted one
file that it complained about. File was:
librocksdb-sys/rocksdb/util/channel.h

Also hacked some dep that I don't remember to avoid using a feature that
was unstable. This was quite dirty but I was sick of dealing with cargo
deps.