validator: Add --wait-for-exit flag to exit subcommand (#6233)
agave-validator exit currents returns immediately after the AdminRpc
call returns. However, the running validator has not exited at this
point and may continue to tear itself down for multiple seconds

The exit subcommand now has an optional flag, --wait-for-exit, that
queries the PID from the validator and loops until that PID has fully
terminated. Use of this flag means that a caller can be sure the
running validator is dead when agave-validator exit returns