validator: Adjust --wait-for-exit message on non-linux (#6708)
The validator exit command can have --monitor OR --wait-for-exit
specified. The functionality for --wait-for-exit is only supported on
linux, and the error message for the non-linux codepath uses the term
"monitor"

Given that the other option is called monitor, this is potentially
confusing. So, adjust the message to use "wait" instead of "monitor"