ledger: Remove threadpool from Shredder
The Shredder threadpool has been around since the days of legacy shreds.
Merkle shreds were adopted on the cluster a while back and with that,
the operations that were fanned out to the Shredder threadpool have been
greatly reduced. The operations are small enough now where using the
threadpool actually adds more overhead than the actual work to be done.

So, remove the threadpool and perform the per-shreds operations with a
single thread